requirements.txt
streamlit
pandas 
numpy
faiss-cpu
sentence-transformers
groq
requests
